Pull csv writing funciton out to a separate module.
